+++ src/test/specs/templateValuesSrv-specs.js
@@ -27,20 +27,6 @@ define([
       });
     });
 
-    describe.only('should init values', function() {
-      var variables = [
-        { name: 'test', current: { value: 'hej' }}
-      ];
-      var dashboard = { templating: { list: variables } };
-
-      beforeEach(function() {
-        ctx.service.init(dashboard);
-      });
-
-      it('should update options array', function() {
-      });
-    });
-
     function describeUpdateVariable(desc, fn) {
       describe(desc, function() {
         var scenario = {};
@@ -139,12 +125,12 @@ define([
     describeUpdateVariable('and existing value still exists in options', function(scenario) {
       scenario.setup(function() {
         scenario.variable = { type: 'query', query: 'apps.*', name: 'test' };
-        scenario.variable.current = { value: 'backend2'};
+        scenario.variable.current = { text: 'backend2'};
         scenario.queryResult = [{text: 'backend1'}, {text: 'backend2'}];
       });
 
       it('should keep variable value', function() {
-        expect(scenario.variable.current.value).to.be('backend2');
+        expect(scenario.variable.current.text).to.be('backend2');
       });
     });
 
@@ -193,18 +179,6 @@ define([
 
       it('should return matches options', function() {
         expect(scenario.variable.options.length).to.be(1);
-      });
-    });
-
-    describeUpdateVariable('and existing value still exists in options', function(scenario) {
-      scenario.setup(function() {
-        scenario.variable = { type: 'query', query: 'apps.*', name: 'test' };
-        scenario.variable.current = { value: 'backend2'};
-        scenario.queryResult = [{text: 'backend1'}, {text: 'backend2'}];
-      });
-
-      it('should keep variable value', function() {
-        expect(scenario.variable.current.value).to.be('backend2');
       });
     });